# web/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Campaign Engine API",
    "description": "Channel-agnostic retail campaign engine. Send the API key in the \u0060X-Api-Key\u0060 header. Polymorphic rules use a \u0060type\u0060 discriminator; see docs/rule-reference.md.",
    "license": {
      "name": "MIT"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/campaigns": {
      "get": {
        "tags": [
          "Campaigns"
        ],
        "summary": "List campaigns",
        "description": "Filters by status, channel (campaigns without channels match every channel), tag, a free-text search on code and name, and \u0060activeFrom\u0060 / \u0060activeTo\u0060: campaigns whose schedule window overlaps that range. \u0060sort\u0060 = updated (default), priority, name, code, startsAt, endsAt; \u0060order\u0060 = desc (default) or asc.",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "channel",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "activeFrom",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "activeTo",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfCampaign"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Create a campaign (always as draft)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Campaign"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Campaign"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{idOrCode}/description": {
      "get": {
        "tags": [
          "Campaigns"
        ],
        "summary": "The campaign in plain English, section by section",
        "parameters": [
          {
            "name": "idOrCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignDescription"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/describe": {
      "post": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Describe an unsaved definition in plain English",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Campaign"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignDescription"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{idOrCode}": {
      "get": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Get a campaign by id or code",
        "parameters": [
          {
            "name": "idOrCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Campaign"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{id}": {
      "put": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Replace a campaign definition",
        "description": "Send the \u0060version\u0060 you loaded; a different stored version returns 409. Live campaigns are re-checked for conflicts.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "force",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Campaign"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignChange"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Delete a draft campaign",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/campaigns/{id}/activate": {
      "post": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Activate a draft or paused campaign",
        "description": "Returns the conflicts with other live campaigns. Error-level conflicts (e.g. duplicate coupon codes) block activation unless force=true.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "force",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignChange"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{id}/pause": {
      "post": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Pause an active campaign",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Campaign"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{id}/archive": {
      "post": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Archive a campaign (terminal)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Campaign"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/validate": {
      "post": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Validate a definition without saving it",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Campaign"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/conflicts": {
      "post": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Conflicts of a (possibly unsaved) campaign with live campaigns",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Campaign"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CampaignConflict"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Campaigns"
        ],
        "summary": "All conflicts among live campaigns",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CampaignConflict"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/ledger/usage/campaigns": {
      "get": {
        "tags": [
          "Ledger"
        ],
        "summary": "Redemptions, discount and limit / budget consumption per campaign",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CampaignUsage"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ledger/usage/campaigns/{campaignId}": {
      "get": {
        "tags": [
          "Ledger"
        ],
        "summary": "Usage of one campaign",
        "parameters": [
          {
            "name": "campaignId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignUsage"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/analytics/campaigns": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Performance of every redeemed or active campaign, best first",
        "description": "\u0060from\u0060 (inclusive) and \u0060to\u0060 (exclusive) are ISO 8601; the default is the last 30 days, the maximum 366 days. Reversed transactions are excluded.",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CampaignPerformance"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Campaign": {
        "required": [
          "code",
          "name",
          "reward"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string",
            "description": "Unique, human-friendly code. Printed on receipts and used by other systems."
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/CampaignStatus"
          },
          "priority": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Higher values are applied first and win ties.",
            "format": "int32"
          },
          "stacking": {
            "$ref": "#/components/schemas/StackingMode"
          },
          "exclusivityGroup": {
            "type": [
              "null",
              "string"
            ],
            "description": "At most one campaign per group applies to a cart: the one giving the largest discount."
          },
          "currency": {
            "type": [
              "null",
              "string"
            ],
            "description": "If set, the campaign only applies to carts in this currency."
          },
          "schedule": {
            "$ref": "#/components/schemas/Schedule"
          },
          "channels": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Channels the campaign runs on. Empty means every channel."
          },
          "stores": {
            "$ref": "#/components/schemas/StoreScope"
          },
          "customerSegments": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Customer segments the campaign is for. Empty means everyone, including anonymous customers."
          },
          "coupon": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "description": "When set, one of the coupon codes must be present in the cart.",
                "$ref": "#/components/schemas/CouponRule"
              }
            ]
          },
          "target": {
            "description": "Products the reward applies to (and that count toward conditions by default).",
            "$ref": "#/components/schemas/ProductSelector"
          },
          "conditions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Condition"
            },
            "description": "All conditions must hold. Use \u0060anyOf\u0060 / \u0060not\u0060 for other logic."
          },
          "reward": {
            "$ref": "#/components/schemas/Reward"
          },
          "limits": {
            "$ref": "#/components/schemas/CampaignLimits"
          },
          "ignoreGlobalExclusions": {
            "type": "boolean",
            "description": "Allows discounting products in global exclusion lists. Use with care."
          },
          "displayMessage": {
            "type": [
              "null",
              "string"
            ],
            "description": "Text shown to the customer / cashier when the campaign applies."
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Free-form references to other systems, e.g. \u0060{\u0022erpCampaignNo\u0022: \u00222026-0042\u0022}\u0060."
          },
          "version": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Incremented on every change; used for optimistic concurrency.",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "description": "A campaign definition: who gets (audience), what (target \u002B reward), when (schedule),\nunder which circumstances (conditions) and how often (limits)."
      },
      "CampaignChange": {
        "required": [
          "campaign",
          "conflicts",
          "warnings"
        ],
        "type": "object",
        "properties": {
          "campaign": {
            "$ref": "#/components/schemas/Campaign"
          },
          "conflicts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CampaignConflict"
            }
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "CampaignConflict": {
        "required": [
          "campaignId",
          "campaignCode",
          "otherId",
          "otherCode",
          "kind",
          "severity",
          "certainty",
          "message"
        ],
        "type": "object",
        "properties": {
          "campaignId": {
            "type": "string",
            "format": "uuid"
          },
          "campaignCode": {
            "type": "string"
          },
          "otherId": {
            "type": "string",
            "format": "uuid"
          },
          "otherCode": {
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/ConflictKind"
          },
          "severity": {
            "$ref": "#/components/schemas/ConflictSeverity"
          },
          "certainty": {
            "$ref": "#/components/schemas/OverlapCertainty"
          },
          "message": {
            "type": "string"
          },
          "combinedRateEstimate": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d\u002B)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
      "CampaignDescription": {
        "required": [
          "summary",
          "reward",
          "conditions",
          "products",
          "audience",
          "schedule",
          "limits",
          "combination"
        ],
        "type": "object",
        "properties": {
          "summary": {
            "type": "string"
          },
          "reward": {
            "type": "string"
          },
          "conditions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "products": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "audience": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "schedule": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "limits": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "combination": {
            "type": "string"
          }
        },
        "description": "A campaign definition in plain English, section by section."
      },
...
      "CampaignPerformance": {
        "required": [
          "campaignId",
          "code",
          "name",
          "status",
          "redemptions",
          "discount",
          "averageDiscount",
          "discountShare",
          "budget",
          "budgetUsed",
          "maxRedemptions"
        ],
        "type": "object",
        "properties": {
          "campaignId": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/CampaignStatus"
          },
          "redemptions": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "discount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d\u002B)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "averageDiscount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d\u002B)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "discountShare": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d\u002B)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "budget": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d\u002B)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "budgetUsed": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d\u002B)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "maxRedemptions": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "CampaignStatus": {
        "enum": [
          "draft",
          "active",
          "paused",
          "archived"
        ]
      },
      "CampaignUsage": {
        "required": [
          "campaignId",
          "code",
          "name",
          "status",
          "redemptions",
          "discountTotal",
          "maxRedemptions",
          "budget",
          "budgetRemaining",
          "reversedRedemptions",
          "lastRedeemedAt"
        ],
        "type": "object",
        "properties": {
          "campaignId": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/CampaignStatus"
          },
          "redemptions": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "discountTotal": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d\u002B)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "maxRedemptions": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "budget": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d\u002B)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "budgetRemaining": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d\u002B)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "reversedRedemptions": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "lastRedeemedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
...
      "PagedResultOfCampaign": {
        "required": [
          "items",
          "page",
          "pageSize",
          "totalCount"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Campaign"
            }
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "StackingMode": {
        "enum": [
          "stackable",
          "exclusive"
        ]
      },
...
      "ValidationResponse": {
        "required": [
          "valid",
          "errors",
          "warnings"
        ],
        "type": "object",
        "properties": {
          "valid": {
            "type": "boolean"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "description": "API key of the calling system.",
        "name": "X-Api-Key",
        "in": "header"
      },
      "Session": {
        "type": "http",
        "description": "Session token from POST /api/v1/auth/login (web app users).",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "ApiKey": []
    }
  ],
...
}